+++ src/components/TextareaField/TextareaField.test.tsx
@@ -1,7 +1,7 @@
 import { generateSnapshots } from '@chanzuckerberg/story-utils';
 import type { StoryFile } from '@storybook/testing-react';
 
-import { render, screen } from '@testing-library/react';
+import { act, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import React from 'react';
@@ -49,5 +49,57 @@ describe('<TextareaField />', () => {
     await user.keyboard('{arrowdown}{delete}');
 
     expect(onChangeFn).not.toHaveBeenCalled();
+  });
+
+  it('will not fire a custom event when at max length', async () => {
+    const onChangeFn = jest.fn();
+    const user = userEvent.setup();
+    render(
+      <TextareaField
+        aria-label="test"
+        defaultValue="1234567890"
+        maxLength={10}
+        onChange={() => onChangeFn()}
+      />,
+    );
+
+    const field = screen.getByRole('textbox');
+
+    expect(field).not.toHaveFocus();
+
+    await userEvent.tab();
+
+    expect(field).toHaveFocus();
+
+    await user.keyboard('abc');
+
+    expect(onChangeFn).not.toHaveBeenCalled();
+  });
+
+  it('will fire a custom event when at max recommended length', async () => {
+    const onChangeFn = jest.fn();
+    const user = userEvent.setup();
+    render(
+      <TextareaField
+        aria-label="test"
+        defaultValue="1234567890"
+        onChange={() => onChangeFn()}
+        recommendedMaxLength={10}
+      />,
+    );
+
+    const field = screen.getByRole('textbox');
+
+    expect(field).not.toHaveFocus();
+
+    await userEvent.tab();
+
+    expect(field).toHaveFocus();
+
+    await act(async () => {
+      await user.keyboard('abc');
+    });
+
+    expect(onChangeFn).toHaveBeenCalled();
   });
 });